+++ src/pages/Timeline/index.ts
@@ -1,1 +1,2 @@
-export { default } from "./Timeline";
+export { Timeline } from "./Timeline";
+export {DefaultTimeline} from './DefaultTimeline';
